pyproject.toml
[project]
name = "django-healthrecordapi"
version = "0.1.0"
description = "A secure and efficient API for managing patient health records"
readme = "README.md"
requires-python = ">=3.11.5"
dependencies = [
    "django>=5.2.1",
    "djangorestframework>=3.14.0",
    "psycopg2-binary>=2.9.9",
    "gunicorn>=21.2.0",
    "python-jose>=3.3.0",
    "django-cors-headers>=4.3.1",
    "django-filter>=23.5",
    "django-environ>=0.11.2",
    "dotenv>=0.9.9",
    "rest-framework-simplejwt>=0.0.2",
    "djangorestframework-simplejwt==5.3.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-django>=4.11.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["healthrecords"]
